backend/requirements.txt
asgiref==3.6.0
black==23.3.0
click==8.1.3
Django==4.0.10
django-cors-headers==3.10.1
djangorestframework==3.14.0
flake8==6.0.0
gunicorn==20.1.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2023.3
sqlparse==0.4.3
tomli==2.0.1
whitenoise==5.3.0
